imprimir comanda

+++ cuenta.h
@@ -10,6 +10,6 @@ void AtenderMesa(int *cuentas[],int mesa,t_producto p[],int totalP,t_categoria c
 int MostrarProductosxCategoria (t_producto p[],int totalP,int numCat,t_categoria c[],int totalCat);
 int getProducto(t_producto p[],int totalP,int numCat,t_categoria c[],int totalCat, int opcion);
 void printCuenta (int *cuentas[],int mesa);
-int getPrecio(t_producto p[],int totalP,int numCat,t_categoria c[],int totalCat, int opcion);
+float getPrecio(t_producto p[],int totalP,int numCat,t_categoria c[],int totalCat, int opcion);
 
 #endif
